requirements.txt
logging
newspaper4k
unidecode
PyYAML
lxml_html_clean
typing_extensions